+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add support for N+1 detection with prosopite
+
 ## [0.1.3] - 2024-11-14
 
 - Enable allocation profiling by default